% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_arc.R
\name{gg_arc}
\alias{gg_arc}
\title{data.frame for ggplot2 arc line}
\usage{
gg_arc(xmin = 0, xmax = 10, ymin = 0, height = 10, n_pt = 100)
}
\arguments{
\item{xmin}{numeric value specifying minimum horizontal value for arc (default = 0)}

\item{xmax}{numeric value specifying maximum horizontal value for arc (default = 10)}

\item{ymin}{numeric value specifying minimum vertival value for arc (default = 0)}

\item{height}{numeric value specifying height of arc. Negative value can be
specified for arc extending down (default = 10)}

\item{n_pt}{numeric value specifying number of points used to draw arc (default = 100)}
}
\value{
\code{data.frame} with 2 columns specifyin x and y coordinates of points of arc
to be passed to \code{ggplot2::geom_path()}
}
\description{
helper function for adding horizontal arc to ggplot2 object via geom_path()
modified from: http://stackoverflow.com/questions/6862742/
}
\author{
Patrick Kimes
}

