% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_heatmap.R
\name{gg_heatmap}
\alias{gg_heatmap}
\title{plot nice heatmap and dendrograms using ggplot2}
\usage{
gg_heatmap(data, distance, linkage, row_sort = TRUE, labmat = NULL,
  rlabmat = NULL, colbrew = "Set1", rcolbrew = "Set1")
}
\arguments{
\item{data}{a numeric matrix of expression, genes x samples}

\item{distance}{a character string for \code{dist}, or a function which returns
a \code{dist} object}

\item{linkage}{a character string for \code{hclust} to use for clustering}

\item{row_sort}{a logical whether rows should be reordered based on
clustering (default = TRUE)}

\item{colbrew}{a vector of strings same length as \code{labframe} specifying
the colors to be used for column labels, if length is 1 same palette is
repeated, see Details regarding valid strings (default = "Set1")}

\item{rcolbrew}{a vector of strings same length as \code{rlabframe} specifying
       the colors to be used for row labels, if length is 1 same palette is
       repeated, see Details regarding valid strings (default = "Set1")}

\item{labframe}{a data.frame of column labels (default = NULL)}

\item{rlabframe}{a data.frame of row labels (default = NULL)}
}
\value{
a \code{ggplot} object
}
\description{
plot nice heatmap and dendrograms using ggplot2
}
\details{
Valid choices of color palettes are listed in \code{RColorBrewer::brewer.pal.info}
}
\author{
Patrick Kimes
}

