% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_multi.R
\name{gg_multi}
\alias{gg_multi}
\title{multiple plot function for ggplot2}
\usage{
gg_multi(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{sequence of \code{ggplot} objects}

\item{plotlist}{list of \code{ggplot} objects if \code{...} not passed (default = NULL)}

\item{cols}{numeric value specifying number of columns in layout (default = 1)}

\item{layout}{matrix specifying the layout, if present \code{cols} is
       ignored (default = NULL)}
}
\description{
Function is reporduced from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\details{
If the layout is something like matrix(c(1, 2, 3, 3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\author{
Winston Chang (http://www.cookbook-r.com/)
}

