% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gg_pcaplot.R
\name{gg_pcaplot}
\alias{gg_pcaplot}
\title{pca scores plot in ggplot2}
\usage{
gg_pcaplot(x, npc, labs = NULL, idiag = TRUE)
}
\arguments{
\item{x}{a matrix of data (n x p)}

\item{npc}{an integer number of principal components}

\item{labs}{labels for the different clusters}

\item{idiag}{a boolean whether to include}
}
\value{
a \code{ggpairs} plot of the PC scores
}
\description{
pca scores plot in ggplot2
}
\author{
Patrick Kimes
}

