% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggheatmap.R
\name{ggheatmap}
\alias{ggheatmap}
\title{Heatmaps Using ggplot2

This function can be used to create heatmaps with dendrograms similar to \code{heatmap.2}
using \code{ggplot2} and \code{ggdendro} for a more "modern" and clean appearance.}
\usage{
ggheatmap(data, distance = "euclidean", linkage = "average",
  row_sort = TRUE, col_sort = TRUE, labframe = NULL, rlabframe = NULL,
  colbrew = "Set1", rcolbrew = "Set1", rtext = NULL, fill_title = NULL)
}
\arguments{
\item{data}{a numeric matrix of expression, genes x samples, or a \code{dist} object}

\item{distance}{a character string for \code{dist}, or a function which returns
a \code{dist} object (default = "euclidean")}

\item{linkage}{a character string for \code{hclust} to use for clustering
(default = "average")}

\item{row_sort}{a logical whether rows should be reordered based on
clustering (default = TRUE)}

\item{col_sort}{a logical whether columns should be reordered based on
clustering (default = TRUE)}

\item{labframe}{a data.frame of column labels (default = NULL)}

\item{rlabframe}{a data.frame of row labels (default = NULL)}

\item{colbrew}{a vector of strings same length as \code{labframe} specifying
the colors to be used for column labels, if length is 1 same palette is
repeated, see Details regarding valid strings (default = "Set1")}

\item{rcolbrew}{a vector of strings same length as \code{rlabframe} specifying
the colors to be used for row labels, if length is 1 same palette is
repeated, see Details regarding valid strings (default = "Set1")}

\item{rtext}{a vector of strings with text annotations for rows of heatmap
(default = NULL)}

\item{fill_title}{a character string specifying the fill title (default = NULL)}
}
\value{
a \code{ggplot} object
}
\description{
Heatmaps Using ggplot2

This function can be used to create heatmaps with dendrograms similar to \code{heatmap.2}
using \code{ggplot2} and \code{ggdendro} for a more "modern" and clean appearance.
}
\details{
Valid choices of color palettes are listed in \code{RColorBrewer::brewer.pal.info}
}
\author{
Patrick Kimes
}

