% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggmulti.R
\name{ggmulti}
\alias{ggmulti}
\title{Plot Multiple ggplot2 Objects

Function is reproduced from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
\usage{
ggmulti(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{sequence of \code{ggplot} objects}

\item{plotlist}{list of \code{ggplot} objects if \code{...} not passed (default = NULL)}

\item{cols}{numeric value specifying number of columns in layout (default = 1)}

\item{layout}{matrix specifying the layout, if present \code{cols} is
       ignored (default = NULL)}
}
\description{
Plot Multiple ggplot2 Objects

Function is reproduced from http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
\details{
If the layout is something like matrix(c(1, 2, 3, 3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\author{
Winston Chang (http://www.cookbook-r.com/)
}

