% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ggpcaplot.R
\name{ggpcaplot}
\alias{ggpcaplot}
\title{PCA Scores Plot Using ggplot2

This function is a simple wrapper for plotting two-way scatter plots
of principal component scores using \code{GGally::ggpairs}.}
\usage{
ggpcaplot(x, npc, labs = NULL, idiag = TRUE)
}
\arguments{
\item{x}{a matrix of data (n x p)}

\item{npc}{an integer number of principal components}

\item{labs}{labels for the different clusters}

\item{idiag}{a boolean whether to include}
}
\value{
a \code{ggpairs} plot of the PC scores
}
\description{
PCA Scores Plot Using ggplot2

This function is a simple wrapper for plotting two-way scatter plots
of principal component scores using \code{GGally::ggpairs}.
}
\author{
Patrick Kimes
}

