% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/color_conv.R
\name{hex2rgb}
\alias{hex2rgb}
\title{HEX string to RGB triple}
\usage{
hex2rgb(hexcolors)
}
\arguments{
\item{hexcolors}{vector of HEX color strings}
}
\value{
matrix of RGB colors with three columns
}
\description{
helper function for converting HEX colors to RGB colors
}
\author{
Patrick Kimes
}

