% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkpackages.R
\name{pkpackages}
\alias{pkpackages}
\title{Create R Package for Patrick

This function is modeled after Hadley's \code{devtools::create} and other similar functions
for defining and producing consistent packages using templates. The function is written to several
different types of R packages/directories specified in the Details.}
\usage{
pkpackages(path, type = "method")
}
\arguments{
\item{path}{string specifying location and name of package}

\item{type}{string specifying type of package to create. Must be one of a set of pre-defined
       types else function returns error (default = "analysis")}
}
\description{
Create R Package for Patrick

This function is modeled after Hadley's \code{devtools::create} and other similar functions
for defining and producing consistent packages using templates. The function is written to several
different types of R packages/directories specified in the Details.
}
\details{
Currently supported valid types include:
\itemize{
\item "method": for a new method/algorithm, closest to standard R package
\item "sim": for method evaluation using simulations
\item "data": for method evaluation using real data examples
\item "analysis": for analyses run independent of method testing
}
}
\author{
Patrick Kimes
}

