% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/randrot.R
\name{randrot}
\alias{randrot}
\title{Random rotation matrix}
\usage{
randrot(p = 2)
}
\arguments{
\item{p}{dimension of desired random rotation matrix (default = 2)}
}
\value{
p x p random rotation matrix
}
\description{
Function for generating a random rotation matrix in R^p
(uses QR decomposition of random normal distribution)
}
\examples{
rotated_x <- tcrossprod(x, RandRot(ncol(x))) #nxp matrix
}
\author{
Patrick Kimes
}

