% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randrotate.R
\name{randrotate}
\alias{randrotate}
\title{Random rotation matrix

Function for generating a random rotation matrix in R^p
using the QR decomposition of random normal distribution.}
\usage{
randrotate(p = 2)
}
\arguments{
\item{p}{dimension of desired random rotation matrix (default = 2)}
}
\value{
p x p random rotation matrix
}
\description{
Random rotation matrix

Function for generating a random rotation matrix in R^p
using the QR decomposition of random normal distribution.
}
\examples{
\dontrun{
rotated_x <- tcrossprod(x, randrotate(ncol(x))) #nxp matrix
}
}
\author{
Patrick Kimes
}

