% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/color_conv.R
\name{rgb2hex}
\alias{rgb2hex}
\title{RGB triple to HEX string}
\usage{
rgb2hex(rgb)
}
\arguments{
\item{rgb}{matrix with three columns with entries between (0, 255)}
}
\value{
vector of HEX color strings
}
\description{
helper function for converting matrix of RGB colors on (0, 255) scale
to HEX color strings
}
\author{
Patrick Kimes
}

